refactor(cli): extract commands/docs.ts and commands/collections.ts (batch 4)
- commands/docs.ts completes: getDocument/multiGet/listFiles/formatLsTime
  join renderFullPath. multiGet keeps all 6 output formats verbatim.
- commands/collections.ts: initLocalIndex, context add/list/rm,
  collection list/remove/rename, detectCollectionFromPath.
- collectionAdd deliberately stays in qmd.ts until batch 5: it calls
  indexFiles (moving to indexing.ts then), and moving it now would
  create a collections.ts → qmd.ts import cycle (I6).

Verified: tsc clean, build ok, 32 snapshots byte-identical,
cli/lazy-llm-import/exit-lifecycle green, cli.test.ts 146/146.
