+++ src/cli/commands/collections.ts
@@ -1,0 +1,371 @@
+/**
+ * commands/collections.ts - collection CRUD (list/remove/rename + init) and
+ * context CRUD (add/list/remove). collectionAdd lands here in batch 5
+ * together with indexing.ts, which it orchestrates.
+ */
+import { realpathSync, existsSync, mkdirSync } from "fs";
+import { join as pathJoin, resolve as pathResolve } from "path";
+import type { Database } from "../../db.js";
+import {
+  getPwd,
+  getRealPath,
+  homedir,
+  resolve,
+  createStore,
+  syncConfigToDb,
+  listCollections,
+  removeCollection,
+  renameCollection,
+  isVirtualPath,
+  parseVirtualPath,
+} from "../../store.js";
+import { resolveModels } from "../../llm.js";
+import {
+  getCollection as getCollectionFromYaml,
+  listCollections as yamlListCollections,
+  addContext as yamlAddContext,
+  removeContext as yamlRemoveContext,
+  removeCollection as yamlRemoveCollectionFn,
+  renameCollection as yamlRenameCollectionFn,
+  setGlobalContext,
+  listAllContexts,
+  loadConfig,
+  saveConfig,
+  setConfigSource,
+} from "../../collections.js";
+import {
+  getDb,
+  getStore,
+  resyncConfig,
+  closeDb,
+  setStoreDbPathOverride,
+  ensureModelsConfiguredForCli,
+} from "../context.js";
+import { c, formatTimeAgo } from "../term.js";
+
+function sameDirectory(a: string, b: string): boolean {
+  try {
+    return realpathSync(a) === realpathSync(b);
+  } catch {
+    return pathResolve(a) === pathResolve(b);
+  }
+}
+
+function initLocalIndex(): void {
+  const cwd = getPwd();
+  if (sameDirectory(cwd, homedir())) {
+    throw new Error("Refusing to initialize a local index in $HOME. The global index is automatically created; run `qmd collection add <path>` for the global index, or run `qmd init` inside a project folder.");
+  }
+
+  const qmdDir = pathJoin(cwd, ".qmd");
+  const ymlPath = pathJoin(qmdDir, "index.yml");
+  const yamlPath = pathJoin(qmdDir, "index.yaml");
+  const configPath = existsSync(yamlPath) ? yamlPath : ymlPath;
+  const dbPath = pathJoin(qmdDir, "index.sqlite");
+
+  mkdirSync(qmdDir, { recursive: true });
+  setConfigSource({ configPath });
+  setStoreDbPathOverride(dbPath);
+  closeDb();
+
+  if (!existsSync(configPath)) {
+    saveConfig({
+      collections: {},
+      models: resolveModels(),
+    });
+  } else {
+    ensureModelsConfiguredForCli();
+  }
+
+  const localStore = createStore(dbPath);
+  syncConfigToDb(localStore.db, loadConfig());
+  localStore.close();
+
+  console.log("ready to go with new local index");
+}
+
+function detectCollectionFromPath(db: Database, fsPath: string): { collectionName: string; relativePath: string } | null {
+  const realPath = getRealPath(fsPath);
+
+  // Find collections that this path is under from YAML
+  const allCollections = yamlListCollections();
+
+  // Find longest matching path
+  let bestMatch: { name: string; path: string } | null = null;
+  for (const coll of allCollections) {
+    if (realPath.startsWith(coll.path + '/') || realPath === coll.path) {
+      if (!bestMatch || coll.path.length > bestMatch.path.length) {
+        bestMatch = { name: coll.name, path: coll.path };
+      }
+    }
+  }
+
+  if (!bestMatch) return null;
+
+  // Calculate relative path
+  let relativePath = realPath;
+  if (relativePath.startsWith(bestMatch.path + '/')) {
+    relativePath = relativePath.slice(bestMatch.path.length + 1);
+  } else if (relativePath === bestMatch.path) {
+    relativePath = '';
+  }
+
+  return {
+    collectionName: bestMatch.name,
+    relativePath
+  };
+}
+
+async function contextAdd(pathArg: string | undefined, contextText: string): Promise<void> {
+  const db = getDb();
+
+  // Handle "/" as global context (applies to all collections)
+  if (pathArg === '/') {
+    setGlobalContext(contextText);
+    resyncConfig();
+    console.log(`${c.green}✓${c.reset} Set global context`);
+    console.log(`${c.dim}Context: ${contextText}${c.reset}`);
+    closeDb();
+    return;
+  }
+
+  // Resolve path - defaults to current directory if not provided
+  let fsPath = pathArg || '.';
+  if (fsPath === '.' || fsPath === './') {
+    fsPath = getPwd();
+  } else if (fsPath.startsWith('~/')) {
+    fsPath = homedir() + fsPath.slice(1);
+  } else if (!fsPath.startsWith('/') && !fsPath.startsWith('qmd://')) {
+    fsPath = resolve(getPwd(), fsPath);
+  }
+
+  // Handle virtual paths (qmd://collection/path)
+  if (isVirtualPath(fsPath)) {
+    const parsed = parseVirtualPath(fsPath);
+    if (!parsed) {
+      console.error(`${c.yellow}Invalid virtual path: ${fsPath}${c.reset}`);
+      process.exit(1);
+    }
+
+    const coll = getCollectionFromYaml(parsed.collectionName);
+    if (!coll) {
+      console.error(`${c.yellow}Collection not found: ${parsed.collectionName}${c.reset}`);
+      process.exit(1);
+    }
+
+    yamlAddContext(parsed.collectionName, parsed.path, contextText);
+    resyncConfig();
+
+    const displayPath = parsed.path
+      ? `qmd://${parsed.collectionName}/${parsed.path}`
+      : `qmd://${parsed.collectionName}/ (collection root)`;
+    console.log(`${c.green}✓${c.reset} Added context for: ${displayPath}`);
+    console.log(`${c.dim}Context: ${contextText}${c.reset}`);
+    closeDb();
+    return;
+  }
+
+  // Detect collection from filesystem path
+  const detected = detectCollectionFromPath(db, fsPath);
+  if (!detected) {
+    console.error(`${c.yellow}Path is not in any indexed collection: ${fsPath}${c.reset}`);
+    console.error(`${c.dim}Run 'qmd status' to see indexed collections${c.reset}`);
+    process.exit(1);
+  }
+
+  yamlAddContext(detected.collectionName, detected.relativePath, contextText);
+  resyncConfig();
+
+  const displayPath = detected.relativePath ? `qmd://${detected.collectionName}/${detected.relativePath}` : `qmd://${detected.collectionName}/`;
+  console.log(`${c.green}✓${c.reset} Added context for: ${displayPath}`);
+  console.log(`${c.dim}Context: ${contextText}${c.reset}`);
+  closeDb();
+}
+
+function contextList(): void {
+  const db = getDb();
+
+  const allContexts = listAllContexts();
+
+  if (allContexts.length === 0) {
+    console.log(`${c.dim}No contexts configured. Use 'qmd context add' to add one.${c.reset}`);
+    closeDb();
+    return;
+  }
+
+  console.log(`\n${c.bold}Configured Contexts${c.reset}\n`);
+
+  let lastCollection = '';
+  for (const ctx of allContexts) {
+    if (ctx.collection !== lastCollection) {
+      console.log(`${c.cyan}${ctx.collection}${c.reset}`);
+      lastCollection = ctx.collection;
+    }
+
+    const displayPath = ctx.path ? `  ${ctx.path}` : '  / (root)';
+    console.log(`${displayPath}`);
+    console.log(`    ${c.dim}${ctx.context}${c.reset}`);
+  }
+
+  closeDb();
+}
+
+function contextRemove(pathArg: string): void {
+  if (pathArg === '/') {
+    // Remove global context
+    setGlobalContext(undefined);
+    // Resync so SQLite store_config is updated
+    const s = getStore();
+    resyncConfig();
+    closeDb();
+    console.log(`${c.green}✓${c.reset} Removed global context`);
+    return;
+  }
+
+  // Handle virtual paths
+  if (isVirtualPath(pathArg)) {
+    const parsed = parseVirtualPath(pathArg);
+    if (!parsed) {
+      console.error(`${c.yellow}Invalid virtual path: ${pathArg}${c.reset}`);
+      process.exit(1);
+    }
+
+    const coll = getCollectionFromYaml(parsed.collectionName);
+    if (!coll) {
+      console.error(`${c.yellow}Collection not found: ${parsed.collectionName}${c.reset}`);
+      process.exit(1);
+    }
+
+    const success = yamlRemoveContext(coll.name, parsed.path);
+
+    if (!success) {
+      console.error(`${c.yellow}No context found for: ${pathArg}${c.reset}`);
+      process.exit(1);
+    }
+
+    console.log(`${c.green}✓${c.reset} Removed context for: ${pathArg}`);
+    return;
+  }
+
+  // Handle filesystem paths
+  let fsPath = pathArg;
+  if (fsPath === '.' || fsPath === './') {
+    fsPath = getPwd();
+  } else if (fsPath.startsWith('~/')) {
+    fsPath = homedir() + fsPath.slice(1);
+  } else if (!fsPath.startsWith('/')) {
+    fsPath = resolve(getPwd(), fsPath);
+  }
+
+  const db = getDb();
+  const detected = detectCollectionFromPath(db, fsPath);
+  closeDb();
+
+  if (!detected) {
+    console.error(`${c.yellow}Path is not in any indexed collection: ${fsPath}${c.reset}`);
+    process.exit(1);
+  }
+
+  const success = yamlRemoveContext(detected.collectionName, detected.relativePath);
+
+  if (!success) {
+    console.error(`${c.yellow}No context found for: qmd://${detected.collectionName}/${detected.relativePath}${c.reset}`);
+    process.exit(1);
+  }
+
+  console.log(`${c.green}✓${c.reset} Removed context for: qmd://${detected.collectionName}/${detected.relativePath}`);
+}
+
+// Collection management commands
+function collectionList(): void {
+  const db = getDb();
+  const collections = listCollections(db);
+
+  if (collections.length === 0) {
+    console.log("No collections found. Run 'qmd collection add .' to create one.");
+    closeDb();
+    return;
+  }
+
+  console.log(`${c.bold}Collections (${collections.length}):${c.reset}\n`);
+
+  for (const coll of collections) {
+    const updatedAt = coll.last_modified ? new Date(coll.last_modified) : new Date();
+    const timeAgo = formatTimeAgo(updatedAt);
+
+    // Get YAML config to check includeByDefault
+    const yamlColl = getCollectionFromYaml(coll.name);
+    const excluded = yamlColl?.includeByDefault === false;
+    const excludeTag = excluded ? ` ${c.yellow}[excluded]${c.reset}` : '';
+
+    console.log(`${c.cyan}${coll.name}${c.reset} ${c.dim}(qmd://${coll.name}/)${c.reset}${excludeTag}`);
+    console.log(`  ${c.dim}Pattern:${c.reset}  ${coll.glob_pattern}`);
+    if (yamlColl?.ignore?.length) {
+      console.log(`  ${c.dim}Ignore:${c.reset}   ${yamlColl.ignore.join(', ')}`);
+    }
+    console.log(`  ${c.dim}Files:${c.reset}    ${coll.active_count}`);
+    console.log(`  ${c.dim}Updated:${c.reset}  ${timeAgo}`);
+    console.log();
+  }
+
+  closeDb();
+}
+
+function collectionRemove(name: string): void {
+  // Check if collection exists in YAML
+  const coll = getCollectionFromYaml(name);
+  if (!coll) {
+    console.error(`${c.yellow}Collection not found: ${name}${c.reset}`);
+    console.error(`Run 'qmd collection list' to see available collections.`);
+    process.exit(1);
+  }
+
+  const db = getDb();
+  const result = removeCollection(db, name);
+  // Also remove from YAML config
+  yamlRemoveCollectionFn(name);
+  closeDb();
+
+  console.log(`${c.green}✓${c.reset} Removed collection '${name}'`);
+  console.log(`  Deleted ${result.deletedDocs} documents`);
+  if (result.cleanedHashes > 0) {
+    console.log(`  Cleaned up ${result.cleanedHashes} orphaned content hashes`);
+  }
+}
+
+function collectionRename(oldName: string, newName: string): void {
+  // Check if old collection exists in YAML
+  const coll = getCollectionFromYaml(oldName);
+  if (!coll) {
+    console.error(`${c.yellow}Collection not found: ${oldName}${c.reset}`);
+    console.error(`Run 'qmd collection list' to see available collections.`);
+    process.exit(1);
+  }
+
+  // Check if new name already exists in YAML
+  const existing = getCollectionFromYaml(newName);
+  if (existing) {
+    console.error(`${c.yellow}Collection name already exists: ${newName}${c.reset}`);
+    console.error(`Choose a different name or remove the existing collection first.`);
+    process.exit(1);
+  }
+
+  const db = getDb();
+  renameCollection(db, oldName, newName);
+  // Also rename in YAML config
+  yamlRenameCollectionFn(oldName, newName);
+  closeDb();
+
+  console.log(`${c.green}✓${c.reset} Renamed collection '${oldName}' to '${newName}'`);
+  console.log(`  Virtual paths updated: ${c.cyan}qmd://${oldName}/${c.reset} → ${c.cyan}qmd://${newName}/${c.reset}`);
+}
+
+export {
+  initLocalIndex,
+  contextAdd,
+  contextList,
+  contextRemove,
+  collectionList,
+  collectionRemove,
+  collectionRename,
+};
