+++ src/cli/qmd.ts
@@ -14,13 +14,10 @@ import {
   enableProductionMode,
   searchFTS,
   getContextForFile,
-  getContextForPath,
   listCollections,
   removeCollection,
   renameCollection,
   findSimilarFiles,
-  findDocument,
-  matchFilesByGlob,
   getHashesNeedingEmbedding,
   clearAllEmbeddings,
   insertEmbedding,
@@ -33,8 +30,6 @@ import {
   setCachedResult,
   getIndexHealth,
   parseVirtualPath,
-  buildVirtualPath,
-  isVirtualPath,
   resolveVirtualPath,
   toVirtualPath,
   insertContent,
@@ -56,7 +51,6 @@ import {
   hybridQuery,
   vectorSearchQuery,
   structuredSearch,
-  addLineNumbers,
   type ExpandedQuery,
   DEFAULT_EMBED_MODEL,
   DEFAULT_EMBED_MAX_BATCH_BYTES,
@@ -67,35 +61,23 @@ import {
   createStore,
   reindexCollection,
   generateEmbeddings,
-  syncConfigToDb,
-  findDocumentRef,
-  getDocumentHash,
-  getDocumentContent,
-  countDocumentsInCollection,
-  listDocumentsWithMeta,
   type ReindexResult,
   type ChunkStrategy,
 } from "../store.js";
-import { disposeDefaultLlamaCpp, withLLMSession, pullModels, DEFAULT_MODEL_CACHE_DIR, resolveEmbedModel, resolveGenerateModel, resolveRerankModel, resolveModels } from "../llm.js";
+import { disposeDefaultLlamaCpp, withLLMSession, pullModels, DEFAULT_MODEL_CACHE_DIR, resolveEmbedModel, resolveGenerateModel, resolveRerankModel } from "../llm.js";
 import {
   formatSearchResults,
   formatDocuments,
-  escapeXml,
   type OutputFormat,
 } from "./formatter.js";
 import {
   getCollection as getCollectionFromYaml,
   listCollections as yamlListCollections,
   getDefaultCollectionNames,
-  addContext as yamlAddContext,
-  removeContext as yamlRemoveContext,
   removeCollection as yamlRemoveCollectionFn,
   renameCollection as yamlRenameCollectionFn,
-  setGlobalContext,
   listAllContexts,
   setConfigIndexName,
-  loadConfig,
-  saveConfig,
   setConfigSource,
   findLocalConfigPath,
   getLocalDbPath,
@@ -122,7 +104,6 @@ import {
   isTTY,
   progress,
   formatETA,
-  formatTimeAgo,
   formatMs,
   formatBytes,
   renderProgressBar,
@@ -130,7 +111,6 @@ import {
   formatScore,
   formatExplainNumber,
 } from "./term.js";
-import { renderFullPath } from "./commands/docs.js";
 import {
   printEmptySearchResults,
   outputResults,
@@ -145,6 +125,16 @@ import {
 } from "./help.js";
 import { runSkillsCommand, showSkill, installSkill, outputSkillsJson } from "./commands/skills.js";
 import { showStatus, showDoctor, formatCount, shortModelName } from "./commands/doctor.js";
+import { getDocument, multiGet, listFiles } from "./commands/docs.js";
+import {
+  initLocalIndex,
+  contextAdd,
+  contextList,
+  contextRemove,
+  collectionList,
+  collectionRemove,
+  collectionRename,
+} from "./commands/collections.js";
 
 // I4: these remain importable from src/cli/qmd.ts for existing consumers
 // (test/cli.test.ts, cli-exit-lifecycle.test.ts) after moving to context.ts.
@@ -250,47 +240,6 @@ function checkIndexHealth(db: Database, model: string = resolveEmbedModelForCli(
 
 // Compute unique display path for a document
 // Always include at least parent folder + filename, add more parent dirs until unique
-function sameDirectory(a: string, b: string): boolean {
-  try {
-    return realpathSync(a) === realpathSync(b);
-  } catch {
-    return pathResolve(a) === pathResolve(b);
-  }
-}
-
-function initLocalIndex(): void {
-  const cwd = getPwd();
-  if (sameDirectory(cwd, homedir())) {
-    throw new Error("Refusing to initialize a local index in $HOME. The global index is automatically created; run `qmd collection add <path>` for the global index, or run `qmd init` inside a project folder.");
-  }
-
-  const qmdDir = pathJoin(cwd, ".qmd");
-  const ymlPath = pathJoin(qmdDir, "index.yml");
-  const yamlPath = pathJoin(qmdDir, "index.yaml");
-  const configPath = existsSync(yamlPath) ? yamlPath : ymlPath;
-  const dbPath = pathJoin(qmdDir, "index.sqlite");
-
-  mkdirSync(qmdDir, { recursive: true });
-  setConfigSource({ configPath });
-  setStoreDbPathOverride(dbPath);
-  closeDb();
-
-  if (!existsSync(configPath)) {
-    saveConfig({
-      collections: {},
-      models: resolveModels(),
-    });
-  } else {
-    ensureModelsConfiguredForCli();
-  }
-
-  const localStore = createStore(dbPath);
-  syncConfigToDb(localStore.db, loadConfig());
-  localStore.close();
-
-  console.log("ready to go with new local index");
-}
-
 async function updateCollections(): Promise<void> {
   const db = getDb();
   const storeInstance = getStore();
@@ -388,717 +337,6 @@ async function updateCollections(): Promise<void> {
  * Detect which collection (if any) contains the given filesystem path.
  * Returns { collectionId, collectionName, relativePath } or null if not in any collection.
  */
-function detectCollectionFromPath(db: Database, fsPath: string): { collectionName: string; relativePath: string } | null {
-  const realPath = getRealPath(fsPath);
-
-  // Find collections that this path is under from YAML
-  const allCollections = yamlListCollections();
-
-  // Find longest matching path
-  let bestMatch: { name: string; path: string } | null = null;
-  for (const coll of allCollections) {
-    if (realPath.startsWith(coll.path + '/') || realPath === coll.path) {
-      if (!bestMatch || coll.path.length > bestMatch.path.length) {
-        bestMatch = { name: coll.name, path: coll.path };
-      }
-    }
-  }
-
-  if (!bestMatch) return null;
-
-  // Calculate relative path
-  let relativePath = realPath;
-  if (relativePath.startsWith(bestMatch.path + '/')) {
-    relativePath = relativePath.slice(bestMatch.path.length + 1);
-  } else if (relativePath === bestMatch.path) {
-    relativePath = '';
-  }
-
-  return {
-    collectionName: bestMatch.name,
-    relativePath
-  };
-}
-
-async function contextAdd(pathArg: string | undefined, contextText: string): Promise<void> {
-  const db = getDb();
-
-  // Handle "/" as global context (applies to all collections)
-  if (pathArg === '/') {
-    setGlobalContext(contextText);
-    resyncConfig();
-    console.log(`${c.green}✓${c.reset} Set global context`);
-    console.log(`${c.dim}Context: ${contextText}${c.reset}`);
-    closeDb();
-    return;
-  }
-
-  // Resolve path - defaults to current directory if not provided
-  let fsPath = pathArg || '.';
-  if (fsPath === '.' || fsPath === './') {
-    fsPath = getPwd();
-  } else if (fsPath.startsWith('~/')) {
-    fsPath = homedir() + fsPath.slice(1);
-  } else if (!fsPath.startsWith('/') && !fsPath.startsWith('qmd://')) {
-    fsPath = resolve(getPwd(), fsPath);
-  }
-
-  // Handle virtual paths (qmd://collection/path)
-  if (isVirtualPath(fsPath)) {
-    const parsed = parseVirtualPath(fsPath);
-    if (!parsed) {
-      console.error(`${c.yellow}Invalid virtual path: ${fsPath}${c.reset}`);
-      process.exit(1);
-    }
-
-    const coll = getCollectionFromYaml(parsed.collectionName);
-    if (!coll) {
-      console.error(`${c.yellow}Collection not found: ${parsed.collectionName}${c.reset}`);
-      process.exit(1);
-    }
-
-    yamlAddContext(parsed.collectionName, parsed.path, contextText);
-    resyncConfig();
-
-    const displayPath = parsed.path
-      ? `qmd://${parsed.collectionName}/${parsed.path}`
-      : `qmd://${parsed.collectionName}/ (collection root)`;
-    console.log(`${c.green}✓${c.reset} Added context for: ${displayPath}`);
-    console.log(`${c.dim}Context: ${contextText}${c.reset}`);
-    closeDb();
-    return;
-  }
-
-  // Detect collection from filesystem path
-  const detected = detectCollectionFromPath(db, fsPath);
-  if (!detected) {
-    console.error(`${c.yellow}Path is not in any indexed collection: ${fsPath}${c.reset}`);
-    console.error(`${c.dim}Run 'qmd status' to see indexed collections${c.reset}`);
-    process.exit(1);
-  }
-
-  yamlAddContext(detected.collectionName, detected.relativePath, contextText);
-  resyncConfig();
-
-  const displayPath = detected.relativePath ? `qmd://${detected.collectionName}/${detected.relativePath}` : `qmd://${detected.collectionName}/`;
-  console.log(`${c.green}✓${c.reset} Added context for: ${displayPath}`);
-  console.log(`${c.dim}Context: ${contextText}${c.reset}`);
-  closeDb();
-}
-
-function contextList(): void {
-  const db = getDb();
-
-  const allContexts = listAllContexts();
-
-  if (allContexts.length === 0) {
-    console.log(`${c.dim}No contexts configured. Use 'qmd context add' to add one.${c.reset}`);
-    closeDb();
-    return;
-  }
-
-  console.log(`\n${c.bold}Configured Contexts${c.reset}\n`);
-
-  let lastCollection = '';
-  for (const ctx of allContexts) {
-    if (ctx.collection !== lastCollection) {
-      console.log(`${c.cyan}${ctx.collection}${c.reset}`);
-      lastCollection = ctx.collection;
-    }
-
-    const displayPath = ctx.path ? `  ${ctx.path}` : '  / (root)';
-    console.log(`${displayPath}`);
-    console.log(`    ${c.dim}${ctx.context}${c.reset}`);
-  }
-
-  closeDb();
-}
-
-function contextRemove(pathArg: string): void {
-  if (pathArg === '/') {
-    // Remove global context
-    setGlobalContext(undefined);
-    // Resync so SQLite store_config is updated
-    const s = getStore();
-    resyncConfig();
-    closeDb();
-    console.log(`${c.green}✓${c.reset} Removed global context`);
-    return;
-  }
-
-  // Handle virtual paths
-  if (isVirtualPath(pathArg)) {
-    const parsed = parseVirtualPath(pathArg);
-    if (!parsed) {
-      console.error(`${c.yellow}Invalid virtual path: ${pathArg}${c.reset}`);
-      process.exit(1);
-    }
-
-    const coll = getCollectionFromYaml(parsed.collectionName);
-    if (!coll) {
-      console.error(`${c.yellow}Collection not found: ${parsed.collectionName}${c.reset}`);
-      process.exit(1);
-    }
-
-    const success = yamlRemoveContext(coll.name, parsed.path);
-
-    if (!success) {
-      console.error(`${c.yellow}No context found for: ${pathArg}${c.reset}`);
-      process.exit(1);
-    }
-
-    console.log(`${c.green}✓${c.reset} Removed context for: ${pathArg}`);
-    return;
-  }
-
-  // Handle filesystem paths
-  let fsPath = pathArg;
-  if (fsPath === '.' || fsPath === './') {
-    fsPath = getPwd();
-  } else if (fsPath.startsWith('~/')) {
-    fsPath = homedir() + fsPath.slice(1);
-  } else if (!fsPath.startsWith('/')) {
-    fsPath = resolve(getPwd(), fsPath);
-  }
-
-  const db = getDb();
-  const detected = detectCollectionFromPath(db, fsPath);
-  closeDb();
-
-  if (!detected) {
-    console.error(`${c.yellow}Path is not in any indexed collection: ${fsPath}${c.reset}`);
-    process.exit(1);
-  }
-
-  const success = yamlRemoveContext(detected.collectionName, detected.relativePath);
-
-  if (!success) {
-    console.error(`${c.yellow}No context found for: qmd://${detected.collectionName}/${detected.relativePath}${c.reset}`);
-    process.exit(1);
-  }
-
-  console.log(`${c.green}✓${c.reset} Removed context for: qmd://${detected.collectionName}/${detected.relativePath}`);
-}
-
-/**
- * Render an absolute filesystem path for human display under --full-path.
- *
- * If the path is the current working directory or a subpath of it, return a
- * "./"-prefixed relative path so it is unambiguously a filesystem path (not a
- * bare collection-relative string that could be confused for a `qmd://`
- * fragment). Otherwise return the absolute realpath so symlinks resolve
- * consistently. Returns `null` if the path could not be normalized — callers
- * fall back to whatever they had before.
- */
-function getDocument(filename: string, fromLine?: number, maxLines?: number, lineNumbers?: boolean, fullPath: boolean = false): void {
-  // Parse :line suffix from filename. Two forms:
-  //   "file.md:100"     -> start at line 100
-  //   "file.md:100:40"  -> start at line 100, read 40 lines
-  // The :// in virtual paths is never matched because we anchor digits to $.
-  // Explicit --from/-l flags always win over values parsed from the path.
-  let inputPath = filename;
-  const rangeMatch = inputPath.match(/:(\d+):(\d+)$/);
-  if (rangeMatch) {
-    if (fromLine === undefined) fromLine = parseInt(rangeMatch[1]!, 10);
-    if (maxLines === undefined) maxLines = parseInt(rangeMatch[2]!, 10);
-    inputPath = inputPath.slice(0, -rangeMatch[0].length);
-  } else {
-    const colonMatch = inputPath.match(/:(\d+)$/);
-    if (colonMatch) {
-      const matched = colonMatch[1];
-      if (matched) {
-        if (fromLine === undefined) fromLine = parseInt(matched, 10);
-        inputPath = inputPath.slice(0, -colonMatch[0].length);
-      }
-    }
-  }
-  if (fromLine !== undefined) fromLine = Math.max(1, fromLine);
-
-  const parsedIndexPath = isVirtualPath(inputPath) ? parseVirtualPath(inputPath) : null;
-  if (parsedIndexPath) {
-    if (parsedIndexPath.indexName) {
-      setIndexName(parsedIndexPath.indexName);
-      setConfigIndexName(parsedIndexPath.indexName);
-    }
-    inputPath = buildVirtualPath(parsedIndexPath.collectionName, parsedIndexPath.path);
-  }
-
-  const db = getDb();
-  const doc = findDocument(db, inputPath, { includeBody: true });
-  if ("error" in doc) {
-    if (doc.error === "excluded_by_ignore") {
-      console.error(`Document is excluded by ignore rule: ${filename}`);
-      console.error(`Collection: ${doc.collection}`);
-      console.error(`Matched path: ${doc.path}`);
-      console.error(`Ignore rule: ${doc.rule}`);
-    } else {
-      console.error(`Document not found: ${filename}`);
-      if (doc.similarFiles.length > 0) {
-        console.error("Similar files:");
-        for (const file of doc.similarFiles) console.error(`  ${file}`);
-      }
-    }
-    closeDb();
-    process.exit(1);
-  }
-
-  // `findDocument` already computes the docid (first 6 hash chars) and the
-  // canonical display path, so we reuse them here instead of a second lookup.
-  const docid = doc.docid;
-  const canonicalPath = `qmd://${doc.displayPath}`;
-
-  // --full-path: show the on-disk path instead of the qmd:// URL + docid, when
-  // the file actually exists. Fall back to the canonical header otherwise.
-  let header: string;
-  if (fullPath) {
-    const fsPath = resolveVirtualPath(db, canonicalPath);
-    if (fsPath && existsSync(fsPath)) {
-      header = renderFullPath(fsPath);
-    } else {
-      header = docid ? `${canonicalPath}  #${docid}` : canonicalPath;
-    }
-  } else {
-    header = docid ? `${canonicalPath}  #${docid}` : canonicalPath;
-  }
-
-  let output = doc.body || "";
-  const startLine = fromLine || 1;
-
-  // Apply line filtering if specified
-  if (fromLine !== undefined || maxLines !== undefined) {
-    const lines = output.split('\n');
-    const start = startLine - 1; // Convert to 0-indexed
-    const end = maxLines !== undefined ? start + maxLines : lines.length;
-    output = lines.slice(start, end).join('\n');
-  }
-
-  // Line numbers are on by default (disable with --no-line-numbers) so the
-  // model can cite exact lines and request follow-up ranges via path:from:count.
-  if (lineNumbers) {
-    output = addLineNumbers(output, startLine);
-  }
-
-  // Header: identify the document (path + docid, or the on-disk path with
-  // --full-path), then optional context.
-  console.log(header);
-  if (doc.context) {
-    console.log(`Folder Context: ${doc.context}`);
-  }
-  console.log("---\n");
-  console.log(output);
-  closeDb();
-}
-
-// Multi-get: fetch multiple documents by glob pattern or comma-separated list
-function multiGet(pattern: string, maxLines?: number, maxBytes: number = DEFAULT_MULTI_GET_MAX_BYTES, format: OutputFormat = "cli", lineNumbers: boolean = true, fullPath: boolean = false): void {
-  const db = getDb();
-
-  // Check if it's a comma-separated list or a glob pattern
-  const isCommaSeparated = pattern.includes(',') && !pattern.includes('*') && !pattern.includes('?') && !pattern.includes('{');
-
-  let files: { filepath: string; displayPath: string; bodyLength: number; collection?: string; path?: string }[];
-
-  if (isCommaSeparated) {
-    // Comma-separated list of files (can be virtual paths or relative paths)
-    const names = pattern.split(',').map(s => s.trim()).filter(Boolean);
-    files = [];
-    for (const name of names) {
-      // Resolve qmd:// exact, path exact, then path suffix (store.findDocumentRef).
-      const doc = findDocumentRef(db, name);
-
-      if (doc) {
-        files.push({
-          filepath: doc.virtual_path,
-          displayPath: doc.virtual_path,
-          bodyLength: doc.body_length,
-          collection: doc.collection,
-          path: doc.path
-        });
-      } else {
-        console.error(`File not found: ${name}`);
-      }
-    }
-  } else {
-    // Glob pattern - matchFilesByGlob now returns virtual paths
-    files = matchFilesByGlob(db, pattern).map(f => ({
-      ...f,
-      collection: undefined,  // Will be fetched later if needed
-      path: undefined
-    }));
-    if (files.length === 0) {
-      console.error(`No files matched pattern: ${pattern}`);
-      closeDb();
-      process.exit(1);
-    }
-  }
-
-  // Collect results for structured output
-  const results: { file: string; displayPath: string; fsPath?: string; docid?: string; title: string; body: string; context: string | null; skipped: boolean; skipReason?: string }[] = [];
-
-  for (const file of files) {
-    // Parse virtual path to get collection info if not already available
-    let collection = file.collection;
-    let path = file.path;
-
-    if (!collection || !path) {
-      const parsed = parseVirtualPath(file.filepath);
-      if (parsed) {
-        collection = parsed.collectionName;
-        path = parsed.path;
-      }
-    }
-
-    // Get context using collection-scoped function
-    const context = collection && path ? getContextForPath(db, collection, path) : null;
-
-    // Resolve docid (first 6 chars of content hash) so every entry can be cited.
-    const docHash = collection && path ? getDocumentHash(db, collection, path) : null;
-    const docid = docHash ? docHash.slice(0, 6) : undefined;
-
-    // --full-path: resolve the on-disk path when it exists (else fall back).
-    // Display as ./-prefixed relative path when under $PWD; absolute realpath
-    // otherwise. See renderFullPath() for the policy.
-    let fsPath: string | undefined;
-    if (fullPath) {
-      const resolved = resolveVirtualPath(db, file.filepath);
-      if (resolved && existsSync(resolved)) fsPath = renderFullPath(resolved);
-    }
-
-    // Check size limit
-    if (file.bodyLength > maxBytes) {
-      results.push({
-        file: file.filepath,
-        displayPath: file.displayPath,
-        fsPath,
-        docid,
-        title: file.displayPath.split('/').pop() || file.displayPath,
-        body: "",
-        context,
-        skipped: true,
-        skipReason: `File too large (${Math.round(file.bodyLength / 1024)}KB > ${Math.round(maxBytes / 1024)}KB). Use 'qmd get ${file.displayPath}' to retrieve.`,
-      });
-      continue;
-    }
-
-    // Fetch document content using collection and path
-    if (!collection || !path) continue;
-
-    const doc = getDocumentContent(db, collection, path);
-
-    if (!doc) continue;
-
-    let body = doc.body;
-
-    // Apply line limit if specified
-    if (maxLines !== undefined) {
-      const lines = body.split('\n');
-      body = lines.slice(0, maxLines).join('\n');
-      if (lines.length > maxLines) {
-        body += `\n\n[... truncated ${lines.length - maxLines} more lines]`;
-      }
-    }
-
-    // Line numbers on by default (disable with --no-line-numbers).
-    if (lineNumbers) {
-      body = addLineNumbers(body);
-    }
-
-    results.push({
-      file: file.filepath,
-      displayPath: file.displayPath,
-      fsPath,
-      docid,
-      title: doc.title || file.displayPath.split('/').pop() || file.displayPath,
-      body,
-      context,
-      skipped: false,
-    });
-  }
-
-  closeDb();
-
-  // --full-path replaces the qmd:// path + docid with the on-disk path (when it
-  // resolved). Per result: pick the identifier and whether to show the docid.
-  const identOf = (r: typeof results[number]): string => (fullPath && r.fsPath) ? r.fsPath : r.displayPath;
-  const docidOf = (r: typeof results[number]): string | undefined => (fullPath && r.fsPath) ? undefined : r.docid;
-
-  // Output based on format
-  if (format === "json") {
-    const output = results.map(r => {
-      const docidVal = docidOf(r);
-      return {
-        file: identOf(r),
-        ...(docidVal && { docid: `#${docidVal}` }),
-        title: r.title,
-        ...(r.context && { context: r.context }),
-        ...(r.skipped ? { skipped: true, reason: r.skipReason } : { body: r.body }),
-      };
-    });
-    console.log(JSON.stringify(output, null, 2));
-  } else if (format === "csv") {
-    const escapeField = (val: string | null | undefined): string => {
-      if (val === null || val === undefined) return "";
-      const str = String(val);
-      if (str.includes(",") || str.includes('"') || str.includes("\n")) {
-        return `"${str.replace(/"/g, '""')}"`;
-      }
-      return str;
-    };
-    console.log("docid,file,title,context,skipped,body");
-    for (const r of results) {
-      const docidVal = docidOf(r);
-      console.log([docidVal ? `#${docidVal}` : "", identOf(r), r.title, r.context, r.skipped ? "true" : "false", r.skipped ? r.skipReason : r.body].map(escapeField).join(","));
-    }
-  } else if (format === "files") {
-    for (const r of results) {
-      const docidVal = docidOf(r);
-      const id = docidVal ? `#${docidVal} ` : "";
-      const ctx = r.context ? `,"${r.context.replace(/"/g, '""')}"` : "";
-      const status = r.skipped ? "[SKIPPED]" : "";
-      console.log(`${id}${identOf(r)}${ctx}${status ? `,${status}` : ""}`);
-    }
-  } else if (format === "md") {
-    for (const r of results) {
-      const docidVal = docidOf(r);
-      console.log(`## ${identOf(r)}\n`);
-      if (docidVal) console.log(`**docid:** \`#${docidVal}\`\n`);
-      if (r.title && r.title !== r.displayPath) console.log(`**Title:** ${r.title}\n`);
-      if (r.context) console.log(`**Context:** ${r.context}\n`);
-      if (r.skipped) {
-        console.log(`> ${r.skipReason}\n`);
-      } else {
-        console.log("```");
-        console.log(r.body);
-        console.log("```\n");
-      }
-    }
-  } else if (format === "xml") {
-    console.log('<?xml version="1.0" encoding="UTF-8"?>');
-    console.log("<documents>");
-    for (const r of results) {
-      const docidVal = docidOf(r);
-      const docidAttr = docidVal ? ` docid="#${docidVal}"` : "";
-      console.log(`  <document${docidAttr}>`);
-      console.log(`    <file>${escapeXml(identOf(r))}</file>`);
-      console.log(`    <title>${escapeXml(r.title)}</title>`);
-      if (r.context) console.log(`    <context>${escapeXml(r.context)}</context>`);
-      if (r.skipped) {
-        console.log(`    <skipped>true</skipped>`);
-        console.log(`    <reason>${escapeXml(r.skipReason || "")}</reason>`);
-      } else {
-        console.log(`    <body>${escapeXml(r.body)}</body>`);
-      }
-      console.log("  </document>");
-    }
-    console.log("</documents>");
-  } else {
-    // CLI format (default)
-    for (const r of results) {
-      const docidVal = docidOf(r);
-      const id = docidVal ? `  #${docidVal}` : "";
-      console.log(`\n${'='.repeat(60)}`);
-      console.log(`File: ${identOf(r)}${id}`);
-      console.log(`${'='.repeat(60)}\n`);
-
-      if (r.skipped) {
-        console.log(`[SKIPPED: ${r.skipReason}]`);
-        continue;
-      }
-
-      if (r.context) {
-        console.log(`Folder Context: ${r.context}\n---\n`);
-      }
-      console.log(r.body);
-    }
-  }
-}
-
-// List files in virtual file tree
-function listFiles(pathArg?: string): void {
-  const db = getDb();
-
-  if (!pathArg) {
-    // No argument - list all collections
-    const yamlCollections = yamlListCollections();
-
-    if (yamlCollections.length === 0) {
-      console.log("No collections found. Run 'qmd collection add .' to index files.");
-      closeDb();
-      return;
-    }
-
-    // Get file counts from database for each collection (YAML iteration kept so
-    // zero-document collections still show "0 files").
-    const collections = yamlCollections.map(coll => {
-      return {
-        name: coll.name,
-        file_count: countDocumentsInCollection(db, coll.name)
-      };
-    });
-
-    console.log(`${c.bold}Collections:${c.reset}\n`);
-    for (const coll of collections) {
-      console.log(`  ${c.dim}qmd://${c.reset}${c.cyan}${coll.name}/${c.reset}  ${c.dim}(${coll.file_count} files)${c.reset}`);
-    }
-    closeDb();
-    return;
-  }
-
-  // Parse the path argument
-  let collectionName: string;
-  let pathPrefix: string | null = null;
-
-  const afterScheme = pathArg.startsWith('qmd://') ? pathArg.slice('qmd://'.length) : null;
-  if (afterScheme !== null && afterScheme.startsWith('/')) {
-    // Absolute-path collection: qmd:///Users/foo/bar — normalizeVirtualPath would corrupt
-    // this by stripping all leading slashes, so bypass parseVirtualPath entirely.
-    const normalized = afterScheme.replace(/\/$/, '');
-    const allColls = yamlListCollections();
-    const match = allColls
-      .filter(c => normalized === c.name || normalized.startsWith(c.name + '/'))
-      .sort((a, b) => b.name.length - a.name.length)[0];
-    if (match) {
-      collectionName = match.name;
-      const rest = normalized.slice(match.name.length).replace(/^\//, '');
-      pathPrefix = rest || null;
-    } else {
-      // Preserve the historical qmd:////collection/path alias behavior for normal
-      // collections when no absolute-path collection matches.
-      const parsed = parseVirtualPath(pathArg);
-      if (!parsed) {
-        console.error(`Invalid virtual path: ${pathArg}`);
-        closeDb();
-        process.exit(1);
-      }
-      collectionName = parsed.collectionName;
-      pathPrefix = parsed.path;
-    }
-  } else if (afterScheme !== null) {
-    // Normal virtual path: qmd://collection-name/path
-    const parsed = parseVirtualPath(pathArg);
-    if (!parsed) {
-      console.error(`Invalid virtual path: ${pathArg}`);
-      closeDb();
-      process.exit(1);
-    }
-    collectionName = parsed.collectionName;
-    pathPrefix = parsed.path;
-  } else if (pathArg.startsWith('/')) {
-    // Raw absolute filesystem path — longest-prefix match against collection names
-    const normalized = pathArg.replace(/\/$/, '');
-    const allColls = yamlListCollections();
-    const match = allColls
-      .filter(c => normalized === c.name || normalized.startsWith(c.name + '/'))
-      .sort((a, b) => b.name.length - a.name.length)[0];
-    if (match) {
-      collectionName = match.name;
-      const rest = normalized.slice(match.name.length).replace(/^\//, '');
-      pathPrefix = rest || null;
-    } else {
-      collectionName = normalized;
-    }
-  } else {
-    // Short collection name or name/path
-    const parts = pathArg.split('/');
-    collectionName = parts[0] || '';
-    if (parts.length > 1) {
-      pathPrefix = parts.slice(1).join('/');
-    }
-  }
-
-  // Get the collection
-  const coll = getCollectionFromYaml(collectionName);
-  if (!coll) {
-    console.error(`Collection not found: ${collectionName}`);
-    console.error(`Run 'qmd ls' to see available collections.`);
-    closeDb();
-    process.exit(1);
-  }
-
-  // List files in the collection with size and modification time
-  const files = listDocumentsWithMeta(db, coll.name, pathPrefix ?? undefined);
-
-  if (files.length === 0) {
-    if (pathPrefix) {
-      console.log(`No files found under qmd://${collectionName}/${pathPrefix}`);
-    } else {
-      console.log(`No files found in collection: ${collectionName}`);
-    }
-    closeDb();
-    return;
-  }
-
-  // Calculate max widths for alignment
-  const maxSize = Math.max(...files.map(f => formatBytes(f.size).length));
-
-  // Output in ls -l style
-  for (const file of files) {
-    const sizeStr = formatBytes(file.size).padStart(maxSize);
-    const date = new Date(file.modified_at);
-    const timeStr = formatLsTime(date);
-
-    // Dim the qmd:// prefix, highlight the filename
-    console.log(`${sizeStr}  ${timeStr}  ${c.dim}qmd://${collectionName}/${c.reset}${c.cyan}${file.path}${c.reset}`);
-  }
-
-  closeDb();
-}
-
-// Format date/time like ls -l
-function formatLsTime(date: Date): string {
-  const now = new Date();
-  const sixMonthsAgo = new Date(now.getTime() - 6 * 30 * 24 * 60 * 60 * 1000);
-
-  const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
-  const month = months[date.getMonth()];
-  const day = date.getDate().toString().padStart(2, ' ');
-
-  // If file is older than 6 months, show year instead of time
-  if (date < sixMonthsAgo) {
-    const year = date.getFullYear();
-    return `${month} ${day}  ${year}`;
-  } else {
-    const hours = date.getHours().toString().padStart(2, '0');
-    const minutes = date.getMinutes().toString().padStart(2, '0');
-    return `${month} ${day} ${hours}:${minutes}`;
-  }
-}
-
-// Collection management commands
-function collectionList(): void {
-  const db = getDb();
-  const collections = listCollections(db);
-
-  if (collections.length === 0) {
-    console.log("No collections found. Run 'qmd collection add .' to create one.");
-    closeDb();
-    return;
-  }
-
-  console.log(`${c.bold}Collections (${collections.length}):${c.reset}\n`);
-
-  for (const coll of collections) {
-    const updatedAt = coll.last_modified ? new Date(coll.last_modified) : new Date();
-    const timeAgo = formatTimeAgo(updatedAt);
-    
-    // Get YAML config to check includeByDefault
-    const yamlColl = getCollectionFromYaml(coll.name);
-    const excluded = yamlColl?.includeByDefault === false;
-    const excludeTag = excluded ? ` ${c.yellow}[excluded]${c.reset}` : '';
-
-    console.log(`${c.cyan}${coll.name}${c.reset} ${c.dim}(qmd://${coll.name}/)${c.reset}${excludeTag}`);
-    console.log(`  ${c.dim}Pattern:${c.reset}  ${coll.glob_pattern}`);
-    if (yamlColl?.ignore?.length) {
-      console.log(`  ${c.dim}Ignore:${c.reset}   ${yamlColl.ignore.join(', ')}`);
-    }
-    console.log(`  ${c.dim}Files:${c.reset}    ${coll.active_count}`);
-    console.log(`  ${c.dim}Updated:${c.reset}  ${timeAgo}`);
-    console.log();
-  }
-
-  closeDb();
-}
-
 async function collectionAdd(pwd: string, globPattern: string, name?: string): Promise<void> {
   // If name not provided, generate from pwd basename
   let collName = name;
@@ -1137,55 +375,6 @@ async function collectionAdd(pwd: string, globPattern: string, name?: string): P
   const newColl = getCollectionFromYaml(collName);
   await indexFiles(pwd, globPattern, collName, false, newColl?.ignore);
   console.log(`${c.green}✓${c.reset} Collection '${collName}' created successfully`);
-}
-
-function collectionRemove(name: string): void {
-  // Check if collection exists in YAML
-  const coll = getCollectionFromYaml(name);
-  if (!coll) {
-    console.error(`${c.yellow}Collection not found: ${name}${c.reset}`);
-    console.error(`Run 'qmd collection list' to see available collections.`);
-    process.exit(1);
-  }
-
-  const db = getDb();
-  const result = removeCollection(db, name);
-  // Also remove from YAML config
-  yamlRemoveCollectionFn(name);
-  closeDb();
-
-  console.log(`${c.green}✓${c.reset} Removed collection '${name}'`);
-  console.log(`  Deleted ${result.deletedDocs} documents`);
-  if (result.cleanedHashes > 0) {
-    console.log(`  Cleaned up ${result.cleanedHashes} orphaned content hashes`);
-  }
-}
-
-function collectionRename(oldName: string, newName: string): void {
-  // Check if old collection exists in YAML
-  const coll = getCollectionFromYaml(oldName);
-  if (!coll) {
-    console.error(`${c.yellow}Collection not found: ${oldName}${c.reset}`);
-    console.error(`Run 'qmd collection list' to see available collections.`);
-    process.exit(1);
-  }
-
-  // Check if new name already exists in YAML
-  const existing = getCollectionFromYaml(newName);
-  if (existing) {
-    console.error(`${c.yellow}Collection name already exists: ${newName}${c.reset}`);
-    console.error(`Choose a different name or remove the existing collection first.`);
-    process.exit(1);
-  }
-
-  const db = getDb();
-  renameCollection(db, oldName, newName);
-  // Also rename in YAML config
-  yamlRenameCollectionFn(oldName, newName);
-  closeDb();
-
-  console.log(`${c.green}✓${c.reset} Renamed collection '${oldName}' to '${newName}'`);
-  console.log(`  Virtual paths updated: ${c.cyan}qmd://${oldName}/${c.reset} → ${c.cyan}qmd://${newName}/${c.reset}`);
 }
 
 async function indexFiles(pwd?: string, globPattern: string = DEFAULT_GLOB, collectionName?: string, suppressEmbedNotice: boolean = false, ignorePatterns?: string[]): Promise<void> {
